% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beepr.R
\name{beep_on_warning}
\alias{beep_on_warning}
\title{Play a short sound if there is a warning message}
\usage{
beep_on_warning(expr, sound = 1)
}
\arguments{
\item{expr}{An expression to be evaluated for errors.}

\item{sound}{character string or number specifying what sound to be played by
either specifying one of the built in sounds, specifying the path to a wav
file or specifying an url. The default is 1. Possible sounds are:
\enumerate{ \item \code{"ping"} \item \code{"coin"} \item \code{"fanfare"}
\item \code{"complete"} \item \code{"treasure"} \item \code{"ready"} \item
\code{"shotgun"} \item \code{"mario"} \item \code{"wilhelm"} \item
\code{"facebook"} \item \code{"sword"} } If \code{sound} does not match any
of the sounds above, or is a valid path or url, a random sound will be
played. Currently \code{beep} can only handle http urls, https is not
supported.}
}
\description{
\code{beep_on_warning} wraps an expression and plays a short sound whenever a warning occurs.
}
\details{
If \code{beep} is not able to play the sound a warning is issued rather than
an error. This is in order to not risk aborting or stopping the process that
you wanted to get notified about.
}
\examples{
# Play a "ping" sound when \\code{expr} signals a warning
# In the case below, it plays one every 2 seconds
beep_on_warning({
  warning("First warning")
  Sys.sleep(2)
  warning("Second warning")
  Sys.sleep(2)
  warning("Third warning")
  "end"
})

# Stay silent if \\code{expr} does not produce an error
beep_on_warning(log(1))

\dontrun{
# Play the Wilhelm scream instead of a ping on a warning.
beep_on_warning(warning("oops"), "wilhelm")
}

}
