% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beepr2-package.R
\docType{package}
\name{beepr2}
\alias{beepr2}
\alias{beepr2-package}
\title{Even more Notification Sounds R}
\description{
This package contains one function, beep(), with one purpose: To make it easy
to play notification sounds on whatever platform you are on. It is intended to
be useful, for example, if you are running a long analysis in the background
and want to know when it is ready.
}
\details{
The package just contains one function \code{\link{beep}}, check it out to
see what it does. For sound on Windows and MacOS \pkg{beepr2} depends on the
\pkg{audio} package. For sound on Linux \pkg{beepr2} depends on that either
the paplay utility from the Pulse Audio system, the aplay utility from the
ALSA system, or VLC media player (http://www.videolan.org/vlc/index.html) is
installed and on the PATH. Chances are that you alread have one of these.

The first 11 sounds are the original sounds from the beepr package by Rasmus Bååth.
The sounds starting from index 12 are new in this version including many nostalgic
90s gaming sound fragment. A selection was made for sound clips suitable for end of
script notifcations.

For suggestions for new sounds, feel free to send me an email with the .wav attached
and I can consider adding it in the next version of beepr2.

}
\examples{
# Play a "ping" sound
beep()
}
\author{
Sjors van Heuveln < sjorsvanheuveln@gmail.com >
}

